Refactor: move ResetSeeds inside SeedChip component
